adding the authentication and registration from terminal

+++ hello.sh
@@ -66,13 +66,19 @@ case $1 in
                                 in_array $image_extension
                                 if [[ $? -eq 1 ]] 
                                 then 
-                                        echo "the image $image_name with extension $image_extension will be added "
-                                        fileName=$image_name.$image_extension
-                                        echo $fileName
                                         cd ..
-                                        cp $fileName ./images
-                                        sh ./hello.sh
-
+                                        if [ -f $imageFromUser ]
+                                        then 
+                                                echo "the image $image_name with extension $image_extension will be added "
+                                                fileName=$image_name.$image_extension
+                                                echo $fileName
+                                                cp $fileName ./images
+                                                sh ./hello.sh
+                                        else 
+                                                echo "image not found !"
+                                                echo $imageFromUser
+                                                
+                                        fi 
                                 else 
                                          echo "the image extension is not allowed, only jpeg, png and jpg"
                                 fi
@@ -123,6 +129,92 @@ case $1 in
                         echo "not such directory found !"
                 fi
                 ;;
+        --auth)
+                echo " here u can authenticate or register on the app !"
+                echo "is this your first visit ? still not having an account (y/n)"
+                read response 
+                if [[  $response == "y" ]]
+                then 
+                        echo "you don't have an account, u can register here !"
+                        echo "please enter a username"
+                        read username 
+                        while [[ $username == "" ]] || [[ $username == " " ]]
+                        do
+                                read username
+                        done 
+                         # the username is the only identifier for a user so it must be unique
+                        # verify if the file exists because it's only created when needed  
+                        if [ -f myBase.csv ]
+                        then 
+                        j=0
+                        while grep -q $username myBase.csv 
+                        do
+                                if [[ $j -eq 0 ]]
+                                then 
+                                        echo "username already taken, choose another one"
+                                        read username   
+                                else 
+                                        read username
+                                        j=$((j+1))
+                                fi 
+                        done    
+                        fi
+                        echo "please enter a password, ',' not allowed : "
+                        read -s password 
+                        # verify and validate password 
+                        while [[ $password == "" ]] || [[ $password == " " ]]
+                        do 
+                                read password 
+                        done 
+                        sub=','
+                        if [[ "$password" == *"$sub"* ]]
+                        then
+                                echo "comma not allowed in password, enter password again "
+                                while [[ "$password" == *"$sub"* ]]
+                                do 
+                                        read -s password
+                                done 
+                        fi 
+                        # add data to file.csv
+                        echo $username,$password >> myBase.csv
+                        echo "registration done successefully, now u can authenticate !"
+
+                elif [[  $response == "n" ]]
+                then
+                        echo "okey, you have an account so provide ur crendetials !"
+                        echo "username : "
+                        read username 
+                        while [[ $username == "" ]] || [[ $username == " " ]]
+                        do
+                                read username
+                        done 
+                        if grep -q $username myBase.csv ;
+                        then 
+                        echo -n "password : "
+                        read -s password 
+                        while [[ $password == "" ]] || [[ $password == " " ]]
+                        do 
+                                read password 
+                        done 
+                        # here we have the user input we should only look for a matching in file.csv
+                        grep $username myBase.csv | awk -F ',' '{print $2}' > file.txt 
+                        while read line ;
+                        do 
+                                if [[ $line == $password ]]
+                                then 
+                                echo " Nice to see you again Mr. $username "
+                                 else 
+                                echo " password not matching "
+                                 fi
+                                done < file.txt 
+                                rm file.txt 
+                        else 
+                                echo "no corresponding username found !"
+                        fi 
+                else 
+                        echo "please choose either y or n to respond !"
+                fi 
+        ;;
         *)
                 echo "unkown option, type --help to check all possible commands"
 esac
